apps: react with imports
