+++ pkg/apps/apiserver/registry/rollback/rest_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/openshift/api/apps"
 	appsv1 "github.com/openshift/api/apps/v1"
 	appsfake "github.com/openshift/client-go/apps/clientset/versioned/fake"
+
 	appsapi "github.com/openshift/origin/pkg/apps/apis/apps"
 	_ "github.com/openshift/origin/pkg/apps/apis/apps/install"
 	appsutil "github.com/openshift/origin/pkg/apps/util"
