+++ pkg/apps/controller/deploymentconfig/deploymentconfig_controller_test.go
@@ -23,6 +23,7 @@ import (
 	appsv1 "github.com/openshift/api/apps/v1"
 	appsfake "github.com/openshift/client-go/apps/clientset/versioned/fake"
 	appslisters "github.com/openshift/client-go/apps/listers/apps/v1"
+
 	appsutil "github.com/openshift/origin/pkg/apps/util"
 	appstest "github.com/openshift/origin/pkg/apps/util/test"
 )
