+++ pkg/apps/strategy/util/events.go
@@ -15,6 +15,7 @@ import (
 	"k8s.io/kubernetes/pkg/api/legacyscheme"
 
 	appsutil "github.com/openshift/origin/pkg/apps/util"
+	"github.com/openshift/origin/pkg/apps/util/appsserialization"
 )
 
 // RecordConfigEvent records an event for the deployment config referenced by the
@@ -23,7 +24,7 @@ func RecordConfigEvent(client corev1client.EventsGetter, deployment *corev1.Repl
 	msg string) {
 	t := metav1.Time{Time: time.Now()}
 	var obj runtime.Object = deployment
-	if config, err := appsutil.DecodeDeploymentConfig(deployment); err == nil {
+	if config, err := appsserialization.DecodeDeploymentConfig(deployment); err == nil {
 		obj = config
 	} else {
 		klog.Errorf("Unable to decode deployment config from %s/%s: %v", deployment.Namespace, deployment.Name, err)
