+++ pkg/apps/util/appsserialization/scheme.go
@@ -1,4 +1,4 @@
-package appserialization
+package appsserialization
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
