+++ pkg/apps/util/appsserialization/scheme_test.go
@@ -1,4 +1,4 @@
-package appserialization
+package appsserialization
 
 import (
 	"strings"
